+++ deepgram/clients/listen.py
@@ -2,8 +2,8 @@
 # Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 # SPDX-License-Identifier: MIT
 
-from .prerecorded.client import PreRecordedClient
-from .live.client import LiveClient
+from .prerecorded.client import PreRecordedClient # FUTURE VERSIONINING:, PreRecordedClientV1
+from .live.client import LiveClient # FUTURE VERSIONINING:, LiveClientV1
 from typing import Dict, Any, Optional
 
 
@@ -16,7 +16,17 @@ class ListenClient:
     @property
     def prerecorded(self):
         return PreRecordedClient(self.url, self.headers)
-
+    
+    # FUTURE VERSIONINING:
+    # @property
+    # def prerecorded_v1(self):
+    #     return PreRecordedClientV1(self.url, self.headers)
+    
     @property
     def live(self):
         return LiveClient(self.url, self.api_key, self.headers)
+    
+    # FUTURE VERSIONINING:
+    # @property
+    # def live_v1(self):
+    #     return LiveClientV1(self.url, self.api_key, self.headers)
