+++ deepgram/clients/live/v1_options.py
@@ -4,28 +4,26 @@
 
 from typing import Union, List, TypedDict
 
-class PrerecordedOptions(TypedDict, total=False):
-    alternatives: int
+class LiveOptionsV1(TypedDict, total=False):
     callback: str
-    detect_entities: bool
-    detect_language: bool
-    detect_topics: bool
+    channels: int
     diarize: bool
-    keywords: Union[list, str]
+    encoding: str
+    endpointing: int
+    interim_results: bool
+    keywords: str
     language: str
     model: str
     multichannel: bool
     numerals: bool
-    paragraphs: bool
-    profanity_filter: bool
     punctuate: bool
-    redact: Union[List[str], bool, str]
-    replace: Union[list, str]
-    search: Union[list, str]
+    profanity_filter: bool
+    redact: bool
+    replace: str
+    sample_rate: int
+    search: str
     smart_format: bool
-    summarize: Union[bool, str]
     tag: list
     tier: str
-    utt_split: int
-    utterances: bool
     version: str
+
